Revert "Renamed "ScrivitoExtentions" to "EditorTabs""
This reverts commit 0077748c8d5e885f655b8ba947f41c07dca58cda.
