+++ src/Components/ScrivitoExtensions/IconEditorTab/IconSearchResults.js
