+++ src/Widgets/IconWidget/IconWidgetEditingConfig.js
@@ -36,7 +36,7 @@ Scrivito.provideEditingConfig("IconWidget", {
     {
       title: "Icon",
       component: loadable(() =>
-        import("../../Components/EditorTabs/IconEditorTab")
+        import("../../Components/ScrivitoExtensions/IconEditorTab")
       ),
       key: "icon-group",
     },
